manager refactor
