+++ data/start_tmux.sh
@@ -6,9 +6,8 @@ tmux new-session -d -s flamepaw
 tmux send-keys -t flamepaw ' cd ~/FatesList/modules/infra/flamepaw && ./flamepaw --cmd server; exec $SHELL' Enter
 tmux new-session -d -s manager 
 tmux send-keys -t manager ' flamepaw --cmd site.manager; exec $SHELL' Enter
-tmux new-session -d -s mapleshade
-tmux send-keys -t mapleshade ' cd ~/GitHub-Updates-Bot && source ~/.mapleshadecfg && npm start; exec $SHELL' Enter
 tmux new-session -d -s main 
 tmux send-keys -t main ' flamepaw --cmd site.run; exec $SHELL' Enter
 cpulimit -e flamepaw -l 80 &
 export HISTCONTROL
+echo "Run 'systemd-run --scope -p CPUQuota=50% modules/infra/flamepaw/flamepaw --cmd server' to stop memory leaks"
