--- modules/infra/manager/core/__init__.py
@@ -1,1 +1,0 @@
-from core.core import *
